--- docs/factory-method/boardgame-checkers.css
@@ -1,1 +1,0 @@
-
