+++ docs/factory-method/checkers/checkers_board.dart
@@ -1,4 +1,4 @@
-import 'game.dart';
+import '../game.dart';
 import 'checkers_pieces.dart';
 
 class CheckersBoard implements GameBoard {
