+++ docs/factory-method/chess/boardgame-chess.css
