+++ docs/factory-method/chess/chess_pieces.dart
@@ -1,6 +1,6 @@
 import 'dart:collection';
 import 'dart:html';
-import 'game.dart';
+import '../game.dart';
 
 class ChessPiece extends GamePiece {
   MovementStrategy moveStrategy;
