+++ docs/factory-method/demo.css
@@ -1,7 +1,6 @@
 
 #choices-container {
     background-color: #ccc;
-    padding: 0 10px;
     margin: 10px 0;
     justify-content: center;
     border-radius: 10px;
@@ -18,6 +17,7 @@
     border-radius: 1000px;
     cursor: pointer;
     box-shadow: rgba(0, 0, 0, 0.1) 0px 4px 12px;
+    font-size: 0.8rem;
 }
 
 .game-choice:hover {
@@ -42,7 +42,6 @@
 }
 
 .container {
-    display: flex;
     align-items: center;
 }
