+++ docs/factory-method/index.html
@@ -5,21 +5,17 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     
     <link rel="stylesheet" href="./style.css">
-    <link rel="stylesheet" href="./boardgame-checkers.css">
-    <link rel="stylesheet" href="./boardgame-chess.css">
-    <link rel="stylesheet" href="./boardgame-connect.css">
-    <link rel="stylesheet" href="./boardgame-reversi.css">
+    <link rel="stylesheet" href="./chess/boardgame-chess.css">
+    <link rel="stylesheet" href="./connect/boardgame-connect.css">
+    <link rel="stylesheet" href="./reversi/boardgame-reversi.css">
     <link rel="stylesheet" href="./demo.css">
 
     <title>Factory Method</title>
     
 </head>
 
-
 <body>
-
     <div class="header container bookshelf" id="site-header">
-
         <div class="left-box">
             <div class="anchor-tags">
                 <div class="category-tag"><a href="https://github.com/eeoooue/patterns">Patterns</a></div>
@@ -32,27 +28,14 @@
                 <div class="category-tag"><a href="https://github.com/eeoooue">About</a></div>
             </div>
         </div>
-
-
     </div>
-
-
 
     <div class="" id="demo-area">
-
         <div class="demo-stack totempole">
-                
             <div class="container" id="choices-container"></div>
-
             <div class="container" id="game-container"></div>
-
         </div>
-
     </div>
-
-
-
-
     
     <script src="script.js"></script>
 </body>
