+++ docs/factory-method/reversi/boardgame-reversi.css
