+++ docs/factory-method/script.js.deps
@@ -1,27 +1,27 @@
 file:///C:/tools/dart-sdk/lib/_internal/dart2js_platform.dill
 file:///C:/tools/dart-sdk/lib/libraries.json
-file:///D:/GitHub/Playground/mysite/new-basic/checkers_board.dart
-file:///D:/GitHub/Playground/mysite/new-basic/checkers_game.dart
-file:///D:/GitHub/Playground/mysite/new-basic/checkers_logic.dart
-file:///D:/GitHub/Playground/mysite/new-basic/checkers_pieces.dart
-file:///D:/GitHub/Playground/mysite/new-basic/checkers_view.dart
-file:///D:/GitHub/Playground/mysite/new-basic/chess_board.dart
-file:///D:/GitHub/Playground/mysite/new-basic/chess_game.dart
-file:///D:/GitHub/Playground/mysite/new-basic/chess_logic.dart
-file:///D:/GitHub/Playground/mysite/new-basic/chess_pieces.dart
-file:///D:/GitHub/Playground/mysite/new-basic/chess_view.dart
-file:///D:/GitHub/Playground/mysite/new-basic/connect_board.dart
-file:///D:/GitHub/Playground/mysite/new-basic/connect_game.dart
-file:///D:/GitHub/Playground/mysite/new-basic/connect_logic.dart
-file:///D:/GitHub/Playground/mysite/new-basic/connect_pieces.dart
-file:///D:/GitHub/Playground/mysite/new-basic/connect_view.dart
-file:///D:/GitHub/Playground/mysite/new-basic/game.dart
-file:///D:/GitHub/Playground/mysite/new-basic/reversi_board.dart
-file:///D:/GitHub/Playground/mysite/new-basic/reversi_game.dart
-file:///D:/GitHub/Playground/mysite/new-basic/reversi_logic.dart
-file:///D:/GitHub/Playground/mysite/new-basic/reversi_pieces.dart
-file:///D:/GitHub/Playground/mysite/new-basic/reversi_view.dart
-file:///D:/GitHub/Playground/mysite/new-basic/script.dart
+file:///D:/GitHub/patterns/docs/factory-method/checkers/checkers_board.dart
+file:///D:/GitHub/patterns/docs/factory-method/checkers/checkers_game.dart
+file:///D:/GitHub/patterns/docs/factory-method/checkers/checkers_logic.dart
+file:///D:/GitHub/patterns/docs/factory-method/checkers/checkers_pieces.dart
+file:///D:/GitHub/patterns/docs/factory-method/checkers/checkers_view.dart
+file:///D:/GitHub/patterns/docs/factory-method/chess/chess_board.dart
+file:///D:/GitHub/patterns/docs/factory-method/chess/chess_game.dart
+file:///D:/GitHub/patterns/docs/factory-method/chess/chess_logic.dart
+file:///D:/GitHub/patterns/docs/factory-method/chess/chess_pieces.dart
+file:///D:/GitHub/patterns/docs/factory-method/chess/chess_view.dart
+file:///D:/GitHub/patterns/docs/factory-method/connect/connect_board.dart
+file:///D:/GitHub/patterns/docs/factory-method/connect/connect_game.dart
+file:///D:/GitHub/patterns/docs/factory-method/connect/connect_logic.dart
+file:///D:/GitHub/patterns/docs/factory-method/connect/connect_pieces.dart
+file:///D:/GitHub/patterns/docs/factory-method/connect/connect_view.dart
+file:///D:/GitHub/patterns/docs/factory-method/game.dart
+file:///D:/GitHub/patterns/docs/factory-method/reversi/reversi_board.dart
+file:///D:/GitHub/patterns/docs/factory-method/reversi/reversi_game.dart
+file:///D:/GitHub/patterns/docs/factory-method/reversi/reversi_logic.dart
+file:///D:/GitHub/patterns/docs/factory-method/reversi/reversi_pieces.dart
+file:///D:/GitHub/patterns/docs/factory-method/reversi/reversi_view.dart
+file:///D:/GitHub/patterns/docs/factory-method/script.dart
 org-dartlang-sdk:///lib/_http/crypto.dart
 org-dartlang-sdk:///lib/_http/embedder_config.dart
 org-dartlang-sdk:///lib/_http/http.dart
